+++ Logos/src/utopia/logos/database/storable/url/DomainDbModel.scala
@@ -2,14 +2,15 @@ package utopia.logos.database.storable.url
 
 import utopia.flow.generic.casting.ValueConversions._
 import utopia.flow.generic.model.immutable.Value
+import utopia.flow.util.UncertainBoolean
 import utopia.logos.database.LogosTables
 import utopia.logos.model.factory.url.DomainFactory
 import utopia.logos.model.partial.url.DomainData
 import utopia.logos.model.stored.url.Domain
 import utopia.vault.model.immutable.{DbPropertyDeclaration, Storable}
 import utopia.vault.model.template.HasIdProperty
-import utopia.vault.store.{FromIdFactory, HasId}
 import utopia.vault.nosql.storable.StorableFactory
+import utopia.vault.store.{FromIdFactory, HasId}
 
 import java.time.Instant
 
@@ -36,12 +37,18 @@ object DomainDbModel
 	  */
 	lazy val created = property("created")
 	
+	/**
+	  * Database property used for interacting with is httpses
+	  */
+	lazy val isHttps = property("isHttps")
+	
 	
 	// IMPLEMENTED	--------------------
 	
 	override def table = LogosTables.domain
 	
-	override def apply(data: DomainData) = apply(None, data.url, Some(data.created))
+	override def apply(data: DomainData): DomainDbModel = 
+		apply(None, data.url, Some(data.created), data.isHttps.exact)
 	
 	/**
 	  * @param created Time when this domain was added to the database
@@ -51,9 +58,11 @@ object DomainDbModel
 	
 	override def withId(id: Int) = apply(id = Some(id))
 	
+	override def withIsHttps(isHttps: UncertainBoolean) = apply(isHttps = isHttps.exact)
+	
 	/**
 	  * @param url Full http(s) address of this domain in string format. Includes protocol, 
-	  * domain name and possible port number.
+	  *            domain name and possible port number.
 	  * @return A model containing only the specified url
 	  */
 	override def withUrl(url: String) = apply(url = url)
@@ -67,17 +76,21 @@ object DomainDbModel
   * @author Mikko Hilpinen
   * @since 27.08.2024, v0.3
   */
-case class DomainDbModel(id: Option[Int] = None, url: String = "", created: Option[Instant] = None) 
+case class DomainDbModel(id: Option[Int] = None, url: String = "", created: Option[Instant] = None, 
+	isHttps: Option[Boolean] = None) 
 	extends Storable with HasId[Option[Int]] with FromIdFactory[Int, DomainDbModel] 
 		with DomainFactory[DomainDbModel]
 {
+	// ATTRIBUTES	--------------------
+	
+	override lazy val valueProperties: Seq[(String, Value)] = 
+		Vector(DomainDbModel.id.name -> id, DomainDbModel.url.name -> url, 
+			DomainDbModel.created.name -> created, DomainDbModel.isHttps.name -> isHttps)
+	
+	
 	// IMPLEMENTED	--------------------
 	
 	override def table = DomainDbModel.table
-	
-	override def valueProperties = 
-		Vector(DomainDbModel.id.name -> id, DomainDbModel.url.name -> url, 
-			DomainDbModel.created.name -> created)
 	
 	/**
 	  * @param created Time when this domain was added to the database
@@ -87,9 +100,11 @@ case class DomainDbModel(id: Option[Int] = None, url: String = "", created: Opti
 	
 	override def withId(id: Int) = copy(id = Some(id))
 	
+	override def withIsHttps(isHttps: UncertainBoolean) = copy(isHttps = isHttps.exact)
+	
 	/**
 	  * @param url Full http(s) address of this domain in string format. Includes protocol, 
-	  * domain name and possible port number.
+	  *            domain name and possible port number.
 	  * @return A new copy of this model with the specified url
 	  */
 	override def withUrl(url: String) = copy(url = url)